+++ grid2op/tests/test_Storage.py
@@ -16,6 +16,7 @@ from grid2op.tests.helper_path_test import *
 import grid2op
 from grid2op.Parameters import Parameters
 from grid2op.dtypes import dt_float
+from grid2op.Action import CompleteAction
 
 import warnings
 
@@ -526,3 +527,163 @@ class TestStorageEnv(HelperTests):
         bat_energy_added /= obs.storage_discharging_efficiency  # there are inefficiencies
         # below i said [0., 0.] because i don't have loss on an empty battery
         assert np.all(np.abs(obs.storage_charge - (bat_energy_added + state_of_charge - [0., 0.])) <= self.tol_one)
+
+    def _aux_test_kirchoff(self):
+        p_subs, q_subs, p_bus, q_bus, diff_v_bus = self.env.backend.check_kirchoff()
+        assert np.all(np.abs(p_subs) <= self.tol_one), "error with active value at some substations"
+        assert np.all(np.abs(q_subs) <= self.tol_one), "error with reactive value at some substations"
+        assert np.all(np.abs(p_bus) <= self.tol_one), "error with active value at some bus"
+        assert np.all(np.abs(q_bus) <= self.tol_one), "error with reactive value at some bus"
+        assert np.all(diff_v_bus <= self.tol_one), "error with voltage discrepency"
+
+    def test_storage_action_mw(self):
+        """test the actions are properly implemented in the backend"""
+        array_modif = np.array([-1.5, -10.], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif})
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - array_modif) <= self.tol_one)
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one)
+        self._aux_test_kirchoff()
+
+        array_modif = np.array([2, 8], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif})
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - array_modif) <= self.tol_one)
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one)
+        self._aux_test_kirchoff()
+
+        # illegal action
+        array_modif = np.array([2, 12], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif})
+        obs, reward, done, info = self.env.step(act)
+        assert info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - [0., 0.]) <= self.tol_one)
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one)
+        self._aux_test_kirchoff()
+
+        # full discharge now
+        array_modif = np.array([-1.5, -10.], dtype=dt_float)
+        for nb_ts in range(3):
+            act = self.env.action_space({"set_storage": array_modif})
+            obs, reward, done, info = self.env.step(act)
+            assert not info["exception"]
+            storage_p, storage_q, storage_v = self.env.backend.storages_info()
+            assert np.all(np.abs(storage_p - array_modif) <= self.tol_one), f"error for P for time step {nb_ts}"
+            assert np.all(np.abs(storage_q - 0.) <= self.tol_one), f"error for Q for time step {nb_ts}"
+            self._aux_test_kirchoff()
+
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        # i have emptied second battery
+        assert np.all(np.abs(self.env.backend._grid.storage["p_mw"].values - [-1.5, -4.4599934]) <= self.tol_one)
+        assert np.all(np.abs(obs.storage_charge[1] - 0.) <= self.tol_one)
+        self._aux_test_kirchoff()
+
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        # i have emptied second battery
+        assert np.all(np.abs(self.env.backend._grid.storage["p_mw"].values - [-1.5, 0.]) <= self.tol_one)
+        assert np.all(np.abs(obs.storage_charge[1] - 0.) <= self.tol_one)
+        self._aux_test_kirchoff()
+
+    def test_storage_action_topo(self):
+        """test the modification of the bus of a storage unit"""
+        param = Parameters()
+        param.NB_TIMESTEP_COOLDOWN_SUB = 0
+        param.NB_TIMESTEP_COOLDOWN_LINE = 0
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            env = grid2op.make("educ_case14_storage", test=True, action_class=CompleteAction, param=param)
+        self.env.close()
+        self.env = env
+        obs = self.env.reset()
+
+        # first case, standard modification
+        array_modif = np.array([-1.5, -10.], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif,
+                                     "set_bus": {"storages_id": [(0, 2)],
+                                                 "lines_or_id": [(8, 2)],
+                                                 "generators_id": [(3, 2)]}}
+                                    )
+        obs, reward, done, info = self.env.step(act)
+
+        assert not info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - array_modif) <= self.tol_one)
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one)
+        assert obs.storage_bus[0] == 2
+        assert obs.line_or_bus[8] == 2
+        assert obs.gen_bus[3] == 2
+        self._aux_test_kirchoff()
+
+        # second case, still standard modification (set to orig)
+        array_modif = np.array([1.5, 10.], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif,
+                                     "set_bus": {"storages_id": [(0, 1)],
+                                                 "lines_or_id": [(8, 1)],
+                                                 "generators_id": [(3, 1)]}}
+                                    )
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - array_modif) <= self.tol_one)
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one)
+        assert obs.storage_bus[0] == 1
+        assert obs.line_or_bus[8] == 1
+        assert obs.gen_bus[3] == 1
+        self._aux_test_kirchoff()
+
+        # fourth case: isolated storage on a busbar (so it is disconnected, but with 0. production => so thats fine)
+        array_modif = np.array([0., 7.], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif,
+                                     "set_bus": {"storages_id": [(0, 2)],
+                                                 "lines_or_id": [(8, 1)],
+                                                 "generators_id": [(3, 1)]}}
+                                    )
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - [0., array_modif[1]]) <= self.tol_one), "storage is not disconnected, yet alone on its busbar"
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one)
+        assert obs.storage_bus[0] == -1, "storage should be disconnected"
+        assert storage_v[0] == 0., "storage 0 should be disconnected"
+        assert obs.line_or_bus[8] == 1
+        assert obs.gen_bus[3] == 1
+        self._aux_test_kirchoff()
+
+        # check that if i don't touch it it's set to 0
+        act = self.env.action_space()
+        obs, reward, done, info = self.env.step(act)
+        assert not info["exception"]
+        storage_p, storage_q, storage_v = self.env.backend.storages_info()
+        assert np.all(np.abs(storage_p - 0.) <= self.tol_one), "storage should produce 0"
+        assert np.all(np.abs(storage_q - 0.) <= self.tol_one), "storage should produce 0"
+        assert obs.storage_bus[0] == -1, "storage should be disconnected"
+        assert storage_v[0] == 0., "storage 0 should be disconnected"
+        assert obs.line_or_bus[8] == 1
+        assert obs.gen_bus[3] == 1
+        self._aux_test_kirchoff()
+
+        # trying to act on a disconnected storage => illegal)
+        array_modif = np.array([2., 7.], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif})
+        obs, reward, done, info = self.env.step(act)
+        assert info["exception"]  # action should be illegal
+        assert not done  # this is fine, as it's illegal it's replaced by do nothing
+        self._aux_test_kirchoff()
+
+        # trying to reconnect a storage alone on a bus => game over, not connected bus
+        array_modif = np.array([1., 7.], dtype=dt_float)
+        act = self.env.action_space({"set_storage": array_modif,
+                                     "set_bus": {"storages_id": [(0, 2)],
+                                                 "lines_or_id": [(8, 1)],
+                                                 "generators_id": [(3, 1)]}}
+                                    )
+        obs, reward, done, info = self.env.step(act)
+        assert info["exception"]  # this is a game over
+        assert done